New migrations fix

+++ database/migrations/2024_12_12_054730_add_role_id_to_users_table.php
@@ -14,14 +14,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->foreignId('role_id')
-                ->after('referrer_link')
-                ->references('id')
-                ->on('roles')
-                ->cascadeOnDelete();
+            $table->unsignedBigInteger('role_id')->after('referrer_link')->nullable();
+            $table->foreign('role_id')->references('id')->on('roles')->cascadeOnDelete();
 
             foreach (User::all() as $user) {
-                if ($user->isAdmin()) {
+                if ($user->is_admin) {
                     $user->update(['role_id' => Role::where('code', 'admin')->first()->id]);
                     continue;
                 }
@@ -29,7 +26,9 @@ return new class extends Migration
                 $user->update(['role_id' => Role::where('code', 'subscriber')->first()->id]);
             }
 
-            $table->dropColumn('is_admin');
+            if (Schema::hasColumn('users', 'is_admin')) {
+                $table->dropCo('is_admin');
+            }
         });
     }
 
@@ -42,12 +41,15 @@ return new class extends Migration
             $table->boolean('is_admin')->after('referrer_link')->default(false);
 
             foreach (User::all() as $user) {
-                if ($user->role_id === Role::where('code', 'admin')->first()->id) {
+                if ($user->isAdmin()) {
                     $user->update(['is_admin' => true]);
                 }
             }
 
-            $table->dropColumn('role_id');
+            if (Schema::hasColumn('users', 'role_id')) {
+                $table->dropForeign(['role_id']);
+                $table->dropColumn('role_id');
+            }
         });
     }
 };
